+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 - Compliant fields are recorded and are now part of the Object Standards grade output
-
+- Violations are now recorded and have two types: MISSING and INVALID
 ### Changed
 
 ### Fixed